ecommerce_api/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
transformers==4.38.2
python-dotenv==1.0.0
requests==2.31.0
torch==2.2.0  # Use available version
accelerate==0.27.2  # For model inference acceleration